Fix bug in selecting fs type, resulting in 1K blocksize
One of the previous 64-bit patches (20ce9ade: 64-bit mke2fs cleanup,top
of the 'js/new-64-bit' branch) moved the fs type and feature detection
code to precede the determination of the device size so that we could
check the 64bit flag if the device size would overflow 32 bits.
Unfortunately, the fs type code uses the device size to determine the
default set of parameters, and since the device size wasn't set yet, it
was clearly "small enough" to fit the floppy profile.

The solution is a bit complex.  One problem is that you can't set the
blocks count to > 32 bits if the 64bit feature isn't enabled.  So I
created a function to automatically set the 64bit feature if
necessary.  After we parse the user supplied options, I check if the
64bit feature has been disabled and bail out then.

Signed-off-by: Valerie Aurora Henson <vaurora@redhat.com>
Signed-off-by: "Theodore Ts'o" <tytso@mit.edu>